Add a slight positive contempt for the opponent.
Tuning against Stockfish prefers a negative value here, since Stockfish’ play on average avoids draws.
The rationale for a positive value is that we assume the opponent to favor drawish moves which forces
the search to really find lines which avoid this. This means that against aggressive opponents we
likely lose some amount of visits while converting more games against all weaker ones.