TCEC S21 DivP (#19)
* Update network from JH.94-100 to 69146

* Add a contempt for the playing side.

This lets Leela assume that a draw is slightly bad for the side to move.
Since it is important in DivP to collect enough wins against weaker engines to advance,
this should let it avoid overly drawish positions.

* Add a slight positive contempt for the opponent.

Tuning against Stockfish prefers a negative value here, since Stockfish’ play on average avoids draws.
The rationale for a positive value is that we assume the opponent to favor drawish moves which forces
the search to really find lines which avoid this. This means that against aggressive opponents we
likely lose some amount of visits while converting more games against all weaker ones.

* Increase MovesLeftSlope to 0.0088

This value is a result of a low npm tune against Stockfish. Navs has been using this value in his test run and so far nothing negative was observed.
The evidence base to extrapolate the tune to LTC is poor.

* Use the latest v0.28.0-rc1

* Update update.sh

* Update name

* Decrease MovesLeftMaxEffect from 0.2 to 0.1

Rationale: The goal for MLH is to convert games quickly before the time
runs out. If the MaxEffect is too high, it could end up being detrimental.

* Update LCZero_GPU.json

removed 
```
            {
                "name": "TimeManager",
                "value": "legacy(steepness=7.0)"
            },
```
from the settings, replacing it by tuned smooth TM.

* Decrease RamLimitMb for DivP

* Increase MinibatchSize to 896

* Remove res_block_fusing=true in BackendOptions

This flag is true by default.

* Archive the TCEC S21 DivP configuration

* Add a README.md

* Add benchmarks and backendbences

Co-authored-by: Naphthalin <40385638+Naphthalin@users.noreply.github.com>