TCEC Season 21 Superfinal submission (#20)
List of changes:

* Initialize the TCEC Superfinal submission
* Convert latest submission to symlinks
* Apply updated CPuct settings.
* Increase MaxPrefetch to 896 (equal to MinibatchSize) and set CacheHistoryLength to 1
* Change to sufi21 branch
* Change version in name
* Remove drawscore settings
* Lower MLH slope
* Apply SmartPruningFactor and SmartPruningQEffect
The values are chosen to coincide roughly with the marginal optima.
That means we expect them to work for a wide variety of TM settings.
* Increase RamLimitMb to 96000 (available 128GiB)
* Bump network number from 69146 to 69626
At the time of the commit 69626 is the 3rd place for self-elo and
coincides with the optimum of the moving average.
Should be a robust choice.
* Reset CacheHistoryLength back to 0
* Turn off moves-left-head completely.
* Revert cpuct scaling settings
* Use default values for SPF and SPQE.
These settings have not been validated and we will therefore be conservative.
* Enable SmartPruningQEffect with 0.2 as a very safe value
* Use previous MaxPrefetch of 160
Tests have shown high MaxPrefetch to be an Elo regression.
* Remove nps_update_rate parameter from smooth TM
nps_update_rate was renamed to nps_update_period. Since the parameter was not (and maybe should not be) tuned, it will be removed.